* improvements:
  mxqd: Remove obsolete quickfix for mx_funlock()
  mx_flock: Fix mx_funlock()
  mxqd: Argument to --log-dir has to be a absolute path
  mxqd_control: Cleanup unused functions
  mxqd_control: rename remove_orphaned_group_lists() to server_remove_orphaned_groups()
  mxqd: Add option --no-daemonize
  mxqd: Fix SEGV in server_close when called initialized
  mxqd: Skip loading groups with non-existent users
  mxqd: Make user name mismatch non fatal as long as uname->uid matches
  mxqd/mxqd_control: Move remove_orphaned_group_lists()

+++ mx_flock.c
@@ -133,11 +133,17 @@ int mx_funlock(struct mx_flock *lock)
 {
     int res;
 
-    assert(lock);
+    if (!lock)
+        return 0;
+
+    if (!lock->locked) {
+        _flock_free(lock);
+        return 0;
+    }
+
     assert(lock->fname);
     assert(lock->fd >= 0);
     assert(lock->operation >= 0);
-    assert(lock->locked);
 
     res = unlink(lock->fname);
     if (res < 0)
